+++ trustllm_pkg/trustllm/config.py
@@ -11,3 +11,7 @@ ernie_api = None
 claude_api = None
 palm_api = None
 replicate_api = None
+azure_openai = False
+azure_engine = None
+azure_api_version = "2023-08-01-preview"
+azure_api_base = None